# backend/static/openapi.yml
components:
  schemas:
    Author:
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        scholar_id:
          type: string
          nullable: true
        papers:
          type: array
          readOnly: true
          items:
            type: object
            properties:
              id:
                type: integer
                readOnly: true
              latest_citations:
                type: integer
                readOnly: true
                nullable: true
      type: object
    Citation:
      properties:
        id:
          type: integer
          readOnly: true
        num_cited:
          type: integer
        paper_id:
          type: integer
          readOnly: true
        date:
          type: string
          format: date-time
      type: object
    Paper:
      properties:
        id:
          type: integer
          readOnly: true
        year:
          type: integer
        name:
          type: string
        latest_citation:
          readOnly: true
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Citation'

      type: object
...
info:
  description: API Docs
  title: API Docs
  version: 1.0.0
openapi: 3.0.2
paths:
...
  /api/papers:
    post:
      tags:
        - Papers
      description: Create a new paper
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Paper'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                type: integer
    get:
      tags:
        - Papers
      description: Get all papers
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: page
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          headers:
            Total-Pages:
              description: Total number of pages. Only present if limit is set.
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Paper'
        '404':
          description: Paged too far
  /api/papers/{paper_id}:
    get:
      tags:
        - Papers
      description: Get a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paper'
        '404':
          description: Paper not found
    put:
      tags:
        - Papers
      description: Update a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Paper'
      responses:
        '200':
          description: Updated successfully
        '404':
          description: Paper not found
    delete:
      tags:
        - Papers
      description: Delete a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Paper not found
  /api/papers/{paper_id}/authors:
    get:
      tags:
        - Papers
      description: Get all authors for a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '404':
          description: Paper not found
  /api/papers/{paper_id}/authors/{author_id}:
    put:
      tags:
        - Papers
      description: Add an author to a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Added successfully
        '404':
          description: Author or paper not found
    delete:
      tags:
        - Papers
      description: Remove an author from a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed successfully
        '404':
          description: Author or paper not found
  /api/papers/{paper_id}/citations:
    get:
      tags:
        - Papers
      description: Get all citations for a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citation'
        '404':
          description: Paper not found
    post:
      tags:
        - Papers
      description: Triggers a new citation lookup for a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Citation'
        '404':
          description: Paper not found
        '500':
          description: Error retrieving citations
        '403':
          description: Not enough API credits
  /api/papers/{paper_id}/latest-citations:
    get:
      tags:
        - Papers
      description: Get the latest citations for a paper
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citation'
        '404':
          description: Paper not found
  /api/papers/citations:
    post:
      tags:
        - Papers
      description: Triggers a new citation lookup for a list of paper ids
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citation'
        '500':
          description: Error retrieving citations for one or more papers. Returns citations for all papers that were successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citation'
        '403':
          description: Ran out of API credits. Returns citations for all papers that were successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Citation'
...